Get the interrupted port, fix IOCON register
As by https://www.best-microcontroller-projects.com/mcp23017-interrupt.html
Centipede::portIntPinConfig seems incorrect,
"As identified above bit 0 is an undefined and unused bit. INTPOL should be bit 1, and ODR should be bit 2."

Also the function getIntF has been added to identify the port (device) subject to the interrupt
"""
The library has not been developed far enough for multiple open drain operation and is missing a member function to detect the state of INTFA and INTFB registers 0x0e and 0x0f. These return an interrupt flag state that allows you to determine which port (device) produced an interrupt (and it could be more than one device).
"""